add separate class for extensions to different tags
