+++ lib/src/main/kotlin/com/lemonappdev/konsist/api/ext/list/tagprovider/KoKDocParamTagProviderListExt.kt
@@ -1,0 +1,13 @@
+package com.lemonappdev.konsist.api.ext.list.tagprovider
+
+import com.lemonappdev.konsist.api.declaration.KoKDocTagDeclaration
+import com.lemonappdev.konsist.api.declaration.KoValuedKDocTagDeclaration
+import com.lemonappdev.konsist.api.provider.tag.KoKDocParamTagProvider
+import com.lemonappdev.konsist.api.provider.tag.KoKDocTagsProvider
+
+
+/**
+ * List containing param tags.
+ */
+val <T : KoKDocParamTagProvider> List<T>.paramTags: List<KoValuedKDocTagDeclaration>
+    get() = flatMap { it.paramTags }
