+++ lib/src/main/kotlin/com/lemonappdev/konsist/api/ext/list/tagprovider/KoKDocThrowsTagProviderListExt.kt
@@ -1,0 +1,12 @@
+package com.lemonappdev.konsist.api.ext.list.tagprovider
+
+import com.lemonappdev.konsist.api.declaration.KoKDocTagDeclaration
+import com.lemonappdev.konsist.api.declaration.KoValuedKDocTagDeclaration
+import com.lemonappdev.konsist.api.provider.tag.KoKDocTagsProvider
+import com.lemonappdev.konsist.api.provider.tag.KoKDocThrowsTagProvider
+
+/**
+ * List containing throws tags.
+ */
+val <T : KoKDocThrowsTagProvider> List<T>.throwsTags: List<KoValuedKDocTagDeclaration>
+    get() = flatMap { it.throwsTags }
